requirements.txt
Django==1.11.2
django-solo==1.1.2
django-wysiwyg-redactor==0.5.1
djangorestframework==3.6.3
olefile==0.44
Pillow==4.1.1
pytz==2017.2
